+++ src/Cube.hpp
@@ -6,5 +6,5 @@ class Cube
 {
 	public:
 		static GLfloat cube_vertex[8 * 3];
-		static GLuint cube_elements[12 * 3];
+		static uint_fast8_t cube_elements[12 * 3];
 };
